refactor and fix stopwatch styling
