--- client/src/components/Logout.js
@@ -1,11 +1,0 @@
-import LogoutButton from "./LogoutButton";
-
-export default function Logout() {
-  return (
-    <div className="App">
-      <h1>Logged Out!</h1>
-      <LogoutButton />
-    </div>
-  );
-}
-
