--- client/src/components/Stopwatch.css
@@ -1,9 +1,0 @@
-.stop-watch{
-    height: 85vh;
-    width: 23vw;
-    background-color: #0d0c1b;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-}
